# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QROS — Guest, Staff & Admin API
  version: 1.0.0
  summary: Hợp đồng REST cho hệ thống đặt đồ uống qua QR tại bàn.
  description: |
    Hợp đồng này là **nguồn sự thật** cho ba phía triển khai: Spring Boot (server),
    Next.js (web khách + KDS + quản trị), và các job Python theo lịch. Viết tay trước, sinh code sau —
    không sinh spec từ annotation của backend, vì làm vậy thì backend luôn "đúng" theo định nghĩa
    và hợp đồng mất tác dụng.

    Mỗi operation mang `x-requirements` trỏ về mã yêu cầu trong `PRD-QROS-001` v1.1.

    ## Bốn quy tắc bất biến

    1. **Không có trường giá nào đi lên từ client.** Mọi schema request đặt món đều
       `additionalProperties: false` và không khai báo trường giá. Client gửi `unitPrice`
       sẽ bị chặn ngay ở tầng xác thực lược đồ, trả `400 PRICE_NOT_ACCEPTED` — `ADR-06`.
    2. **Ba vùng, ba chuỗi filter độc lập.** `/guest/**` không bao giờ dùng chung
       cấu hình bảo mật với `/staff/**` hay `/admin/**`.
    3. **Tiền là số nguyên đơn vị đồng.** Không dùng số thực ở bất kỳ đâu.
    4. **Mọi lỗi theo RFC 7807**, kèm phần mở rộng `code` (chuỗi ổn định để máy đọc)
       và `traceId`.
  contact:
    name: QROS Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.qros.example.com
    description: Production
  - url: https://staging-api.qros.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local
...
security: []
paths:
...
  /api/v1/guest/sessions:
    post:
      operationId: startTableSession
      tags: [guest-session]
      summary: Đổi mã QR lấy phiên bàn
      description: |
        Xác minh chữ ký Ed25519 của `qrToken`, kiểm tra bàn tồn tại, chi nhánh đang mở cửa,
        và (nếu bàn bật QR xoay vòng) kiểm TOTP. Trả về phiên bàn gắn với `deviceId`.

        Khi bàn đã có phiên đang hoạt động, hành vi phụ thuộc tình huống — xem `EC-02`:
        thiết bị cùng nhóm được **tham gia** phiên hiện có (`201` với `joined: true`);
        nghi ngờ là nhóm khách mới thì trả `409 TABLE_SESSION_CONFLICT` để chuyển sang nhân viên.
      x-requirements: [FR-CUS-01, NFR-SEC-01, EC-02, EC-03]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartSessionRequest' }
      responses:
        '201':
          description: Phiên đã được tạo hoặc đã tham gia phiên sẵn có.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSession' }
        '401':
          description: '`QR_INVALID_SIGNATURE` · `QR_OTP_EXPIRED`'
          $ref: '#/components/responses/Problem'
        '403':
          description: '`STORE_CLOSED` · bàn ở trạng thái `DISABLED`'
          $ref: '#/components/responses/Problem'
        '409':
          description: '`TABLE_SESSION_CONFLICT` — cần nhân viên xử lý.'
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/guest/sessions/by-code:
    post:
      operationId: startTableSessionByCode
      tags: [guest-session]
      summary: Mở phiên bằng mã bàn 6 ký tự
      description: Đường dự phòng khi camera không quét được QR. Giới hạn 5 lần thử / 10 phút / IP.
      x-requirements: [FR-CUS-02]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartSessionByCodeRequest' }
      responses:
        '201':
          description: Phiên đã được tạo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSession' }
        '401':
          description: '`TABLE_CODE_INVALID`'
          $ref: '#/components/responses/Problem'
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/guest/sessions/current:
    get:
      operationId: getCurrentSession
      tags: [guest-session]
      summary: Trạng thái phiên hiện tại
      description: Trả về phiên bàn gắn với token khách hiện tại nếu phiên còn hiệu lực.
      x-requirements: [FR-CUS-01]
      security: [{ tableSession: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSession' }
        '401':
          description: '`TABLE_SESSION_EXPIRED` — cần quét lại QR.'
          $ref: '#/components/responses/Problem'

  # ────────────────────────────────────────────────────────────────
  #  THỰC ĐƠN
  # ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    tableSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWS `EdDSA`, TTL 90 phút trượt theo hoạt động, gắn với `deviceId`.
        Claim: `iss`, `aud`, `sub`, `exp`, `iat`, `jti`, `sid`, `tid`, `scope`, `tv`.
        Không chứa thông tin cá nhân.
    staffSession:
      type: apiKey
      in: cookie
      name: qros_session
      description: |
        Cookie `HttpOnly; Secure; SameSite=Strict; Path=/api`. TTL 15 phút,
        làm mới qua `/auth/refresh`.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 trên raw body, khoá chia sẻ với cổng thanh toán.
...
  responses:
    Problem:
      description: Lỗi theo RFC 7807.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Vượt hạn mức.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 kèm hai phần mở rộng bắt buộc của QROS.
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: |
            Thông điệp cho người dùng cuối. **Không bao giờ** chứa stack trace,
            câu truy vấn, hay chi tiết nội bộ — `NFR-SEC` A05.
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Mã ổn định để máy đọc, ví dụ `PRICE_NOT_ACCEPTED`.
          examples: [PRICE_NOT_ACCEPTED, ITEM_SOLD_OUT, INVALID_TRANSITION]
        traceId:
          type: string
          description: Tương quan với truy vết OpenTelemetry để nhân viên tra cứu.
...
    StartSessionRequest:
      type: object
      additionalProperties: false
      required: [qrToken, deviceId]
      properties:
        qrToken:
          type: string
          description: JWS Ed25519 đọc từ mã QR trên mặt bàn.
        deviceId:
          type: string
          format: uuid
          description: UUID sinh phía client, lưu lâu dài trong trình duyệt.
        nickname:
          type: string
          maxLength: 24
          description: Biệt danh hiển thị khi nhiều thiết bị chung một bàn.
        coords:
          type: object
          description: Tuỳ chọn, chỉ gửi khi khách cấp quyền vị trí. Chỉ dùng làm tín hiệu phụ.
          additionalProperties: false
          properties:
            lat: { type: number }
            lon: { type: number }
    StartSessionByCodeRequest:
      type: object
      additionalProperties: false
      required: [tableCode, deviceId]
      properties:
        tableCode:
          type: string
          pattern: '^[A-Z0-9]{6}$'
        deviceId: { type: string, format: uuid }
        nickname: { type: string, maxLength: 24 }
    TableSession:
      type: object
      required: [sessionId, storeId, tableId, tableLabel, expiresAt, joined, accessToken]
      properties:
        sessionId: { type: string, format: uuid }
        storeId: { type: string, format: uuid }
        storeName: { type: string }
        tableId: { type: string, format: uuid }
        tableLabel: { type: string, examples: ['A-04'] }
        zoneName: { type: string }
        accessToken:
          type: string
          description: JWS phiên bàn. Client đính kèm ở header `Authorization`.
        expiresAt: { type: string, format: date-time }
        joined:
          type: boolean
          description: '`true` nếu thiết bị tham gia một phiên đã có sẵn — xem `EC-02`.'
        participants:
          type: array
          description: Biệt danh các thiết bị trong phiên. Không chứa dữ liệu cá nhân.
          items:
            type: object
            properties:
              nickname: { type: string }
              isSelf: { type: boolean }
...